feat(map-canvas): add movement and sprite animation utilities
- Introduced movement-animators for handling various NPC movement types including axis-based, wandering, and spinning.
- Implemented sprite-animation utilities to manage sprite frame application and animation updates for markers.
- Added viewport-utils for improved canvas handling, including overscroll computation and effective view size calculation.
- Created an index file to re-export all map canvas utilities for better organization and accessibility.
