+++ maps/bg_palette_metadata.json
@@ -1,5 +1,5 @@
 {
-  "generated_at": "2026-02-14T17:22:42.309286Z",
+  "generated_at": "2026-02-14T17:51:17.155091Z",
   "maps": {
     "AzaleaGym": {
       "label": "AzaleaGym",
