+++ maps/day/animated/map_neighborhoods.json
@@ -1,5 +1,5 @@
 {
-  "generated_at": "2026-02-14T17:22:11.634560Z",
+  "generated_at": "2026-02-14T17:50:43.013323Z",
   "neighborhoods": [
     {
       "bounds_blocks": {
