+++ maps/eve/animated/map_neighborhoods.json
@@ -1,5 +1,5 @@
 {
-  "generated_at": "2026-02-14T17:22:40.497429Z",
+  "generated_at": "2026-02-14T17:51:15.469933Z",
   "neighborhoods": [
     {
       "bounds_blocks": {
