+++ maps/morn/animated/map_neighborhoods.json
@@ -1,5 +1,5 @@
 {
-  "generated_at": "2026-02-14T17:22:21.149427Z",
+  "generated_at": "2026-02-14T17:50:54.187294Z",
   "neighborhoods": [
     {
       "bounds_blocks": {
