+++ maps/nite/animated/map_neighborhoods.json
@@ -1,5 +1,5 @@
 {
-  "generated_at": "2026-02-14T17:22:30.789567Z",
+  "generated_at": "2026-02-14T17:51:04.396122Z",
   "neighborhoods": [
     {
       "bounds_blocks": {
