+++ maps/object_metadata.json
@@ -2060,7 +2060,7 @@
       "label": "Facings"
     }
   },
-  "generated_at": "2026-02-14T17:22:41.822730Z",
+  "generated_at": "2026-02-14T17:51:16.763635Z",
   "maps": {
     "AzaleaGym": {
       "height_blocks": 8,
