+++ maps/warp_metadata.json
@@ -947,7 +947,7 @@
     "YELLOW_FOREST": "YellowForest",
     "YELLOW_FOREST_GATE": "YellowForestGate"
   },
-  "generated_at": "2026-02-14T17:22:40.736254Z",
+  "generated_at": "2026-02-14T17:51:15.636073Z",
   "maps": {
     "AzaleaGym": {
       "collision": {
