+++ maps/weather_state.json
@@ -1,5 +1,5 @@
 {
-  "generated_at": "2026-02-14T17:22:42Z",
+  "generated_at": "2026-02-14T17:51:17Z",
   "weekday": 1,
   "time_of_day": 1,
   "maps": {
