+++ web/atlas/src/components/MapCanvas.tsx
@@ -4,32 +4,23 @@ import {
   Container,
   AnimatedSprite,
   FederatedPointerEvent,
-  Assets,
   Graphics,
   Sprite,
-  Texture,
 } from "pixi.js";
 import { WeatherSystem } from "@/lib/weather";
 import { joinBasePath, withBasePath, withVersion } from "@/lib/basePath";
 import {
   AtlasLayout,
-  MapPlacement,
   MapWarp,
   ObjectMetadata,
-  MapObjectMetadataEntry,
-  ObjectEventEntry,
-  ObjectSpriteDefinition,
   WarpMetadata,
-  MovementSummary,
   BgPalettesMetadata,
 } from "@/types";
 import { registerPixiExtensions } from "@/pixi/registerExtensions";
-import { loadMapAnimation, type MapAnimationResource } from "@/lib/loadMapAnimation";
+import { loadMapAnimation } from "@/lib/loadMapAnimation";
 import { ObjectSpriteCache } from "@/lib/objectSprites";
 import { computeObjectPosition, type PlacementContext } from "@/lib/objectPlacement";
-import { createCollisionHelper, type CollisionHelper } from "@/lib/collision";
-import { getMovementModel } from "@/lib/movementModel";
-import { simulateNpcMovement } from "@/lib/movementSimulation";
+import { createCollisionHelper } from "@/lib/collision";
 import {
   analyzeAllSpriteLimits,
   type SpriteLimitIssue,
@@ -45,112 +36,53 @@ import {
   VIEW_STATE_VERSION
 } from "@/lib/storage";
 
-type OffsetTuple = [number, number];
+// Extracted types
+import type {
+  OffsetTuple,
+  WarpMarkerEntry,
+  WarpBacklink,
+  SpriteFrameRef,
+  MovementFrameSet,
+  SyncedAnimation,
+  OverlayState,
+} from "./MapCanvas/MapCanvas.types";
 
-type WarpMarkerEntry = {
-  warp: MapWarp;
-  graphic: Graphics;
-  // Local pixel offsets within the map sprite
-  localX: number;
-  localY: number;
-};
+// Extracted constants
+import { MIN_SCALE, MAX_SCALE } from "./MapCanvas/constants";
 
-type WarpBacklink = {
-  applicableTo: string | null;
-  mapLabel: string;
-  mapConstant: string | null;
-  warpIndex: number;
-  previous: WarpBacklink | null;
-};
-
-type CardinalDirection = "down" | "up" | "left" | "right";
-
-type Offset = { dx: number; dy: number };
-
-type SpriteFrameRef = {
-  key: string;
-  texture: Texture;
-  offsetX: number;
-  offsetY: number;
-};
-
-type MovementFrameSet = {
-  framesByDirection: Partial<Record<CardinalDirection, SpriteFrameRef[]>>;
-  availableDirections: CardinalDirection[];
-  defaultFrame: SpriteFrameRef;
-  defaultDirection: CardinalDirection | null;
-};
-
-type PokemonIconFrameRecord = NonNullable<
-  ReturnType<ObjectSpriteCache["getPokemonIconFrameTextures"]>
->["frames"][number];
-
-type MovementSegment =
-  | {
-      type: "move";
-      from: Offset;
-      to: Offset;
-      direction: CardinalDirection;
-      durationMs: number;
-      stepIndex: number;
-    }
-  | {
-      type: "wait";
-      position: Offset;
-      direction: CardinalDirection;
-      durationMs: number;
-      stepIndex: number;
-    };
+// Extracted utilities
+import {
+  computeOverscrollPx,
+  computeBottomExtraPx,
+  clampScale,
+  disposeAnimationResource,
+  rendererOn,
+  rendererOff,
+  getEffectiveViewSize,
+  snapToHalf,
+} from "@/lib/map-canvas/viewport-utils";
+import {
+  frameIndexForTime,
+  isObjectVisibleAtTime,
+  resolveFacingConstant,
+  computeMovementSummaryForObject,
+  isObjectWithinMapBounds,
+  resolveDirectionFromFacingKey,
+  buildMovementFrameSet,
+  buildPokemonIconFrameSet,
+} from "@/lib/map-canvas/animation-frames";
+import {
+  createMovementAnimator,
+  createPokemonIconAnimator,
+} from "@/lib/map-canvas/movement-animators";
+import {
+  updateMarkerAnimation,
+} from "@/lib/map-canvas/sprite-animation";
 
 // Local event typings to avoid React typing dependency issues
 type CheckboxChangeEvent = { target: { checked: boolean } };
 type InputNumberChangeEvent = { target: { value: string } };
 type SelectChangeEvent = { target: { value: string } };
-
-type PathMovementAnimator = {
-  kind: "path";
-  segments: MovementSegment[];
-  totalDurationMs: number;
-  stepCount: number;
-};
-
-type SpinStep = {
-  direction: CardinalDirection;
-  durationMs: number;
-};
-
-type SpinMovementAnimator = {
-  kind: "spin";
-  steps: SpinStep[];
-  totalDurationMs: number;
-};
-
-type IdleMovementAnimator = {
-  kind: "idle";
-  direction: CardinalDirection | null;
-  frameCount: number;
-  frameDurationMs: number;
-  phaseOffsetMs: number;
-};
-
-type MovementAnimator = PathMovementAnimator | SpinMovementAnimator | IdleMovementAnimator;
-
-type ObjectMarkerEntry = {
-  object: ObjectEventEntry;
-  sprite: Sprite;
-  movementSummary: MovementSummary | null;
-  animator: MovementAnimator | null;
-  basePosition: { x: number; y: number };
-  spriteOffset: { x: number; y: number };
-  cellPixelSize: number;
-  frameSet: MovementFrameSet | null;
-  currentFrameKey: string | null;
-  spriteScale: number;
-  lastDirection: CardinalDirection | null;
-  currentStepIndex: number | null;
-  stepProgress: number;
-  stepCount: number | null;
-};
 
 interface MapCanvasProps {
   atlas: AtlasLayout | null;
@@ -172,1421 +104,7 @@ interface MapCanvasProps {
   disableObjectAnimations?: boolean;
 }
 
-const MIN_SCALE = 0.25;
-const MAX_SCALE = 4;
-// Allow panning beyond the edge of the content for better UX (esp. on mobile)
-function computeOverscrollPx(viewW: number, viewH: number): number {
-  // 10% of the smaller viewport dimension, clamped to a sensible range
-  const candidate = Math.max(viewW, viewH) > 0 ? Math.min(viewW, viewH) * 0.1 : 64;
-  return Math.max(48, Math.min(256, candidate));
-}
-
-// Provide a slightly larger buffer at the bottom, where browser UI can
-// overlap content and thumbs often need headroom.
-function computeBottomExtraPx(viewH: number): number {
-  // 1.5x the base overscroll or at least 32 additional pixels
-  const base = computeOverscrollPx(viewH, viewH);
-  return Math.max(32, Math.floor(base * 0.5));
-}
-
-
-function clampScale(value: number): number {
-  if (!Number.isFinite(value)) {
-    return 1;
-  }
-  return Math.min(MAX_SCALE, Math.max(MIN_SCALE, value));
-}
-
 registerPixiExtensions();
-
-function disposeAnimationResource(
-  resource: MapAnimationResource | null | undefined,
-  options?: { unload?: boolean },
-): void {
-  if (!resource) {
-    return;
-  }
-  // First, dispose of the Texture wrappers without touching the BaseTexture
-  // so that the asset system can own the BaseTexture lifecycle.
-  for (const texture of resource.textures) {
-    try {
-      // In some teardown orders, PIXI may already have nulled the baseTexture
-      // on a Texture without marking it destroyed. Guard for that to avoid
-      // internal refCount access on a null baseTexture.
-      if (texture && !texture.destroyed) {
-        const bt = (texture as any).baseTexture ?? (texture as any)._baseTexture;
-        if (!bt) {
-          continue;
-        }
-        texture.destroy(false);
-      }
-    } catch {
-      /* ignore individual texture destroy issues */
-    }
-  }
-  // Optionally unload via Assets which will destroy the BaseTexture it manages.
-  // NOTE: Only use unload for short-lived overlay assets; for world map sprites, multiple
-  // instances share the same BaseTexture by URL and unloading can de-texture currently
-  // visible sprites.
-  if (options?.unload) {
-    void Assets.unload(resource.imageUrl);
-  }
-}
-
-function rendererOn(
-  app: Application | null,
-  event: string,
-  handler: (...args: any[]) => void,
-): void {
-  const r: any = app && (app as any).renderer;
-  if (r && typeof r.on === "function") {
-    try {
-      r.on(event, handler);
-    } catch {
-      /* noop */
-    }
-  }
-}
-
-function rendererOff(
-  app: Application | null,
-  event: string,
-  handler: (...args: any[]) => void,
-): void {
-  const r: any = app && (app as any).renderer;
-  if (r && typeof r.off === "function") {
-    try {
-      r.off(event, handler);
-    } catch {
-      /* noop */
-    }
-  }
-}
-
-// Compute the effective visible size of the canvas in CSS pixels, accounting for
-// mobile browser UI that reduces the visual viewport compared to the layout viewport.
-function getEffectiveViewSize(app: Application | null): { width: number; height: number } {
-  if (!app) return { width: 0, height: 0 };
-  const renderer = app.renderer;
-  const canvas = app.view as unknown as HTMLCanvasElement | null;
-  const baseW = Math.max(0, renderer?.width ?? 0);
-  const baseH = Math.max(0, renderer?.height ?? 0);
-  let rectW = baseW;
-  let rectH = baseH;
-  if (canvas) {
-    const rect = canvas.getBoundingClientRect();
-    rectW = Math.max(0, Math.round(rect.width));
-    rectH = Math.max(0, Math.round(rect.height));
-  }
-  let vvW = Number.POSITIVE_INFINITY;
-  let vvH = Number.POSITIVE_INFINITY;
-  if (typeof window !== "undefined" && (window as any).visualViewport) {
-    const vv = window.visualViewport as VisualViewport;
-    vvW = Math.max(0, Math.round(vv.width));
-    vvH = Math.max(0, Math.round(vv.height));
-  }
-  const width = Math.min(baseW || Number.POSITIVE_INFINITY, rectW || Number.POSITIVE_INFINITY, vvW);
-  const height = Math.min(
-    baseH || Number.POSITIVE_INFINITY,
-    rectH || Number.POSITIVE_INFINITY,
-    vvH,
-  );
-  return {
-    width: Number.isFinite(width) ? width : baseW,
-    height: Number.isFinite(height) ? height : baseH,
-  };
-}
-
-type SyncedAnimation = {
-  sprite: AnimatedSprite;
-  resource: MapAnimationResource;
-  placement: MapPlacement;
-  order: number;
-  neighborhoodId: string | null;
-  warpMarkers: WarpMarkerEntry[];
-  objectContainer: Container | null;
-  objectMarkers: ObjectMarkerEntry[];
-  collisionHelper: CollisionHelper | null;
-  // Whether this map sprite is currently within (or near) the viewport
-  // and should be updated/rendered. Used for simple view culling.
-  visible?: boolean;
-  weather?: WeatherSystem | null;
-};
-
-type OverlayState = {
-  mapLabel: string;
-  sprite: AnimatedSprite;
-  resource: MapAnimationResource;
-  background: Graphics;
-  markers: WarpMarkerEntry[];
-  highlight?: Graphics;
-  baseWidth: number;
-  baseHeight: number;
-  cellSize: number;
-  baseAlpha: number;
-  keyHandler: (event: KeyboardEvent) => void;
-  objectContainer: Container | null;
-  objectMarkers: ObjectMarkerEntry[];
-  collisionHelper: CollisionHelper | null;
-  scale: number;
-  fitScale: number;
-  minScale: number;
-  maxScale: number;
-  positioned: boolean;
-  // Sprite limit analysis
-  spriteLimitEnabled?: boolean;
-  spriteIssues?: SpriteLimitIssue[];
-  spriteIssueIndex?: number;
-  spriteIssueHighlight?: Graphics;
-};
-
-function frameIndexForTime(elapsedMs: number, durations: number[], loopDuration: number): number {
-  if (!durations.length || loopDuration <= 0) {
-    return 0;
-  }
-  const cycle = elapsedMs % loopDuration;
-  let acc = 0;
-  for (let index = 0; index < durations.length; index += 1) {
-    acc += durations[index];
-    if (cycle < acc) {
-      return index;
-    }
-  }
-  return durations.length - 1;
-}
-
-function isObjectVisibleAtTime(entry: ObjectEventEntry, timeOfDay: string): boolean {
-  const slots = entry.timeOfDay?.slots;
-  if (!Array.isArray(slots) || slots.length === 0) {
-    return true;
-  }
-  return slots.includes(timeOfDay);
-}
-
-function resolveFacingConstant(entry: ObjectEventEntry, metadata: ObjectMetadata): string | null {
-  const movementKey = entry.movement?.constant ?? "";
-  const movement = movementKey ? metadata.movements[movementKey] : undefined;
-  const movementAction = movement?.action ?? "";
-  if (movementAction === "OBJECT_ACTION_CUT_TREE" && metadata.facings["FACING_CUT_TREE"]) {
-    return "FACING_CUT_TREE";
-  }
-  if (movementAction === "OBJECT_ACTION_FRUIT") {
-    const treeNameRaw = entry.extra?.["tree"];
-    const treeName = typeof treeNameRaw === "string" ? treeNameRaw : "";
-    if (treeName.includes("APRICORN") && metadata.facings["FACING_APRICORN"]) {
-      return "FACING_APRICORN";
-    }
-    if (metadata.facings["FACING_BERRY"]) {
-      return "FACING_BERRY";
-    }
-    if (metadata.facings["FACING_PICKED_FRUIT"]) {
-      return "FACING_PICKED_FRUIT";
-    }
-  }
-  if (movementAction === "OBJECT_ACTION_BIG_GYARADOS") {
-    if (metadata.facings["FACING_BIG_GYARADOS_2"]) {
-      return "FACING_BIG_GYARADOS_2";
-    }
-    if (metadata.facings["FACING_BIG_GYARADOS_1"]) {
-      return "FACING_BIG_GYARADOS_1";
-    }
-  }
-  if (movementAction === "OBJECT_ACTION_BIG_SNORLAX" && metadata.facings["FACING_BIG_DOLL_SYM"]) {
-    return "FACING_BIG_DOLL_SYM";
-  }
-  if (movementAction === "OBJECT_ACTION_SAILBOAT_TOP" && metadata.facings["FACING_SAILBOAT_TOP"]) {
-    return "FACING_SAILBOAT_TOP";
-  }
-  if (
-    movementAction === "OBJECT_ACTION_SAILBOAT_BOTTOM" &&
-    metadata.facings["FACING_SAILBOAT_BOTTOM"]
-  ) {
-    return "FACING_SAILBOAT_BOTTOM";
-  }
-  // Standing flip variants: choose flip facings instead of default.
-  if (movementAction === "OBJECT_ACTION_STAND_FLIP") {
-    const face = (movement?.facing ?? "").toUpperCase();
-    if (face === "DOWN" && metadata.facings["FACING_STEP_DOWN_FLIP"]) {
-      return "FACING_STEP_DOWN_FLIP";
-    }
-    if (face === "UP" && metadata.facings["FACING_STEP_UP_FLIP"]) {
-      return "FACING_STEP_UP_FLIP";
-    }
-  }
-  // Tiny windows use a custom facing series FACING_TINY_WINDOWS_0..6 and the variant
-  // is selected by the object's X range field in map data (args[5] in object_event).
-  // Examples: SnowtopMountainOutside places SPRITE_SAILBOAT with SPRITEMOVEDATA_TINY_WINDOWS
-  // using range.x values 0, 1, 3. Clamp to the defined [0..6] constants and fall back to 0.
-  if (movementAction === "OBJECT_ACTION_TINY_WINDOWS") {
-    const raw = entry.range?.x;
-    const variant = Number.isFinite(raw as number)
-      ? Math.max(0, Math.min(6, Math.trunc(raw as number)))
-      : 0;
-    const key = `FACING_TINY_WINDOWS_${variant}` as const;
-    if (metadata.facings[key]) {
-      return key;
-    }
-    if (metadata.facings["FACING_TINY_WINDOWS_0"]) {
-      return "FACING_TINY_WINDOWS_0";
-    }
-  }
-  const facingValue = movement?.facing ?? "";
-  if (facingValue) {
-    if (metadata.facings[facingValue]) {
-      return facingValue;
-    }
-    const normalized = facingValue.toUpperCase();
-    const mapped =
-      metadata.defaultFacingForDirection[facingValue] ??
-      metadata.defaultFacingForDirection[normalized] ??
-      metadata.defaultFacingForDirection[normalized.toLowerCase()];
-    if (mapped && metadata.facings[mapped]) {
-      return mapped;
-    }
-  }
-  const fallback =
-    metadata.defaultFacingForDirection.DOWN ??
-    metadata.defaultFacingForDirection.Down ??
-    metadata.defaultFacingForDirection.down ??
-    "FACING_STEP_DOWN_0";
-  if (fallback && metadata.facings[fallback]) {
-    return fallback;
-  }
-  const firstKey = Object.keys(metadata.facings)[0];
-  return firstKey ?? null;
-}
-
-function resolvePokemonSpecies(entry: ObjectEventEntry): {
-  species: string | null;
-  form: string | null;
-} {
-  const speciesConstant =
-    typeof entry.species?.constant === "string" ? entry.species.constant : null;
-  const fallbackSpecies =
-    typeof entry.extra?.["species"] === "string" ? (entry.extra!["species"] as string) : null;
-  const rawForm = entry.extra?.["form"];
-  const formConstant = typeof rawForm === "string" ? rawForm : null;
-  return {
-    species: speciesConstant ?? fallbackSpecies,
-    form: formConstant,
-  };
-}
-
-function computeMovementSummaryForObject(
-  objectEntry: ObjectEventEntry,
-  collisionHelper: CollisionHelper | null,
-): MovementSummary | null {
-  const model = getMovementModel(objectEntry.movement?.constant ?? null);
-  try {
-    return simulateNpcMovement({
-      object: objectEntry,
-      model,
-      collisionHelper,
-    });
-  } catch (err) {
-    console.warn("Failed to compute movement summary", objectEntry, err);
-    return null;
-  }
-}
-
-function isObjectWithinMapBounds(
-  objectEntry: ObjectEventEntry,
-  mapData: MapObjectMetadataEntry | null | undefined,
-  cellsPerBlock: number,
-  eventCellPixelSize: number,
-): boolean {
-  if (!mapData) {
-    return true;
-  }
-  const normalisedCellsPerBlock =
-    Number.isFinite(cellsPerBlock) && cellsPerBlock > 0
-      ? Math.trunc(Math.abs(cellsPerBlock))
-      : null;
-  if (!normalisedCellsPerBlock) {
-    return true;
-  }
-  const widthBlocks =
-    Number.isFinite(mapData.widthBlocks) && mapData.widthBlocks && mapData.widthBlocks > 0
-      ? Math.abs(mapData.widthBlocks)
-      : null;
-  const heightBlocks =
-    Number.isFinite(mapData.heightBlocks) && mapData.heightBlocks && mapData.heightBlocks > 0
-      ? Math.abs(mapData.heightBlocks)
-      : null;
-  if (!widthBlocks && !heightBlocks) {
-    return true;
-  }
-  const widthCells = widthBlocks ? widthBlocks * normalisedCellsPerBlock : null;
-  const heightCells = heightBlocks ? heightBlocks * normalisedCellsPerBlock : null;
-  const baseCellSize =
-    Number.isFinite(eventCellPixelSize) && eventCellPixelSize > 0
-      ? Math.abs(eventCellPixelSize)
-      : null;
-
-  const resolveCells = (tiles: number, pixels: number): number | null => {
-    if (Number.isFinite(tiles)) {
-      return tiles;
-    }
-    if (baseCellSize && Number.isFinite(pixels)) {
-      return pixels / baseCellSize;
-    }
-    return null;
-  };
-
-  const xCells = resolveCells(objectEntry.xTiles, objectEntry.xPixels);
-  const yCells = resolveCells(objectEntry.yTiles, objectEntry.yPixels);
-
-  if (widthCells !== null && xCells !== null) {
-    if (xCells < 0 || xCells >= widthCells) {
-      return false;
-    }
-  }
-  if (heightCells !== null && yCells !== null) {
-    if (yCells < 0 || yCells >= heightCells) {
-      return false;
-    }
-  }
-  return true;
-}
-
-const FRAME_DURATION_MS = 1000 / 60;
-const POKEMON_ICON_FRAME_DURATION_SCALE = 2;
-const MIN_POKEMON_ICON_FRAME_DURATION_MS = 120;
-const MOVEMENT_SPEED_SCALE = 2;
-
-const STEP_FRAMES_BY_SPEED: Record<string, number> = {
-  slow: 32,
-  normal: 16,
-  fast: 8,
-};
-
-const IDLE_FRAMES_BY_SPEED: Record<string, number> = {
-  slow: 48,
-  normal: 32,
-  fast: 20,
-};
-
-const SPIN_INTERVAL_BY_SPEED: Record<string, number> = {
-  slow: 700,
-  normal: 540,
-  fast: 360,
-};
-
-const STEP_FACING_KEYS: Record<CardinalDirection, string[]> = {
-  down: ["FACING_STEP_DOWN_0", "FACING_STEP_DOWN_1", "FACING_STEP_DOWN_2", "FACING_STEP_DOWN_3"],
-  up: ["FACING_STEP_UP_0", "FACING_STEP_UP_1", "FACING_STEP_UP_2", "FACING_STEP_UP_3"],
-  left: ["FACING_STEP_LEFT_0", "FACING_STEP_LEFT_1", "FACING_STEP_LEFT_2", "FACING_STEP_LEFT_3"],
-  right: [
-    "FACING_STEP_RIGHT_0",
-    "FACING_STEP_RIGHT_1",
-    "FACING_STEP_RIGHT_2",
-    "FACING_STEP_RIGHT_3",
-  ],
-};
-
-const CLOCKWISE_SEQUENCE: CardinalDirection[] = ["right", "down", "left", "up"];
-const COUNTERCLOCKWISE_SEQUENCE: CardinalDirection[] = ["right", "up", "left", "down"];
-
-const DIRECTION_DELTAS: Record<CardinalDirection, Offset> = {
-  down: { dx: 0, dy: 1 },
-  up: { dx: 0, dy: -1 },
-  left: { dx: -1, dy: 0 },
-  right: { dx: 1, dy: 0 },
-};
-
-function movementSpeedToStepDuration(speed: MovementSummary["model"]["speed"] | undefined): number {
-  const frames = speed ? STEP_FRAMES_BY_SPEED[speed] : undefined;
-  const frameCount = Number.isFinite(frames) ? (frames as number) : STEP_FRAMES_BY_SPEED.normal;
-  return frameCount * FRAME_DURATION_MS * MOVEMENT_SPEED_SCALE;
-}
-
-function movementSpeedToIdleDuration(speed: MovementSummary["model"]["speed"] | undefined): number {
-  const frames = speed ? IDLE_FRAMES_BY_SPEED[speed] : undefined;
-  const frameCount = Number.isFinite(frames) ? (frames as number) : IDLE_FRAMES_BY_SPEED.normal;
-  return frameCount * FRAME_DURATION_MS * MOVEMENT_SPEED_SCALE;
-}
-
-function movementSpeedToSpinInterval(speed: MovementSummary["model"]["speed"] | undefined): number {
-  const interval = speed ? SPIN_INTERVAL_BY_SPEED[speed] : undefined;
-  const base = Number.isFinite(interval) ? (interval as number) : SPIN_INTERVAL_BY_SPEED.normal;
-  return base * MOVEMENT_SPEED_SCALE;
-}
-
-function resolveDirectionFromFacingKey(key: string | null | undefined): CardinalDirection | null {
-  if (!key) {
-    return null;
-  }
-  const normalized = key.toUpperCase();
-  if (normalized.includes("_DOWN")) {
-    return "down";
-  }
-  if (normalized.includes("_UP")) {
-    return "up";
-  }
-  if (normalized.includes("_LEFT")) {
-    return "left";
-  }
-  if (normalized.includes("_RIGHT")) {
-    return "right";
-  }
-  return null;
-}
-
-function createSpriteFrameRef(
-  key: string,
-  record: NonNullable<ReturnType<ObjectSpriteCache["getFacingTexture"]>>,
-): SpriteFrameRef {
-  return {
-    key,
-    texture: record.texture,
-    offsetX: record.offsetX,
-    offsetY: record.offsetY,
-  };
-}
-
-function createSeededRandom(seed: number): () => number {
-  let value = seed >>> 0;
-  return () => {
-    value = (value * 1664525 + 1013904223) >>> 0;
-    return value / 0xffffffff;
-  };
-}
-
-function deltaToDirection(dx: number, dy: number): CardinalDirection | null {
-  if (dx === 0 && dy === 0) {
-    return null;
-  }
-  if (dx === 0) {
-    if (dy > 0) {
-      return "down";
-    }
-    if (dy < 0) {
-      return "up";
-    }
-  }
-  if (dy === 0) {
-    if (dx > 0) {
-      return "right";
-    }
-    if (dx < 0) {
-      return "left";
-    }
-  }
-  return null;
-}
-
-function oppositeDirection(direction: CardinalDirection): CardinalDirection {
-  switch (direction) {
-    case "down":
-      return "up";
-    case "up":
-      return "down";
-    case "left":
-      return "right";
-    case "right":
-      return "left";
-    default:
-      return direction;
-  }
-}
-
-function buildStaticAnimationFrames(
-  cache: ObjectSpriteCache,
-  spriteKey: string,
-  paletteName: string | null,
-  defaultDirection: CardinalDirection | null,
-): { direction: CardinalDirection; frames: SpriteFrameRef[] } | null {
-  if (spriteKey !== "SPRITE_BIG_GYARADOS") {
-    return null;
-  }
-  const facingKeys = ["FACING_BIG_GYARADOS_1", "FACING_BIG_GYARADOS_2"];
-  const frames: SpriteFrameRef[] = [];
-  for (const facingKey of facingKeys) {
-    const record = cache.getFacingTexture(spriteKey, facingKey, paletteName ?? null);
-    if (record) {
-      frames.push(createSpriteFrameRef(facingKey, record));
-    }
-  }
-  if (frames.length === 0) {
-    return null;
-  }
-  if (frames.length === 1) {
-    frames.push(frames[0]);
-  }
-  const direction = defaultDirection ?? "down";
-  return {
-    direction,
-    frames,
-  };
-}
-
-function buildMovementFrameSet(
-  cache: ObjectSpriteCache,
-  spriteKey: string,
-  spriteDef: ObjectSpriteDefinition,
-  paletteName: string | null,
-  baseKey: string,
-  baseRecord: NonNullable<ReturnType<ObjectSpriteCache["getFacingTexture"]>>,
-): MovementFrameSet {
-  const defaultFrame = createSpriteFrameRef(baseKey, baseRecord);
-  const framesByDirection: Partial<Record<CardinalDirection, SpriteFrameRef[]>> = {};
-  const availableDirections: CardinalDirection[] = [];
-  let defaultDirection = resolveDirectionFromFacingKey(baseKey);
-
-  if (spriteDef.spriteType !== "WALKING_SPRITE") {
-    const staticFrames = buildStaticAnimationFrames(
-      cache,
-      spriteKey,
-      paletteName ?? null,
-      defaultDirection ?? null,
-    );
-    if (staticFrames) {
-      framesByDirection[staticFrames.direction] = staticFrames.frames;
-      if (!availableDirections.includes(staticFrames.direction)) {
-        availableDirections.push(staticFrames.direction);
-      }
-      if (!defaultDirection) {
-        defaultDirection = staticFrames.direction;
-      }
-    }
-    return {
-      framesByDirection,
-      availableDirections,
-      defaultFrame,
-      defaultDirection: defaultDirection ?? resolveDirectionFromFacingKey(baseKey),
-    };
-  }
-
-  (Object.keys(STEP_FACING_KEYS) as CardinalDirection[]).forEach((direction) => {
-    const keys = STEP_FACING_KEYS[direction];
-    const frames: SpriteFrameRef[] = [];
-    for (const facingKey of keys) {
-      const record = cache.getFacingTexture(spriteKey, facingKey, paletteName ?? null);
-      if (record) {
-        frames.push(createSpriteFrameRef(facingKey, record));
-      }
-    }
-    if (frames.length > 0) {
-      if (frames.length < 4) {
-        const originals = frames.slice();
-        while (frames.length < 4) {
-          frames.push(originals[frames.length % originals.length]);
-        }
-      }
-      framesByDirection[direction] = frames;
-      if (!availableDirections.includes(direction)) {
-        availableDirections.push(direction);
-      }
-    }
-  });
-
-  defaultDirection = defaultDirection ?? resolveDirectionFromFacingKey(baseKey);
-  if (!defaultDirection && availableDirections.length > 0) {
-    defaultDirection = availableDirections[0];
-  }
-  if (defaultDirection && !availableDirections.includes(defaultDirection)) {
-    availableDirections.unshift(defaultDirection);
-  }
-
-  return {
-    framesByDirection,
-    availableDirections,
-    defaultFrame,
-    defaultDirection: defaultDirection ?? null,
-  };
-}
-
-function buildPokemonIconFrameSet(
-  cache: ObjectSpriteCache,
-  objectEntry: ObjectEventEntry,
-  _spriteDef: ObjectSpriteDefinition,
-  facingKey: string,
-  paletteName: string | null,
-  logContext: string,
-): {
-  baseFrame: SpriteFrameRef | null;
-  frameSet: MovementFrameSet | null;
-  frameDurationMs: number | null;
-} {
-  const spriteKey = "SPRITE_MON_ICON";
-  const { species, form } = resolvePokemonSpecies(objectEntry);
-  const palette = paletteName ?? null;
-  const direction = resolveDirectionFromFacingKey(facingKey) ?? "down";
-  const fallbackRecord = cache.getFacingTexture(spriteKey, facingKey, palette);
-  const iconData = cache.getPokemonIconFrameTextures(species, form, palette);
-
-  const convertFrame = (frameRecord: PokemonIconFrameRecord, index: number): SpriteFrameRef => {
-    const halfWidth = Math.round((frameRecord.width ?? 0) / 2);
-    const halfHeight = Math.round((frameRecord.height ?? 0) / 2);
-    return {
-      key: `${spriteKey}:${species ?? "UNKNOWN"}:${form ?? "NO_FORM"}:${index}`,
-      texture: frameRecord.texture,
-      offsetX: frameRecord.offsetX + halfWidth,
-      offsetY: frameRecord.offsetY + halfHeight,
-    };
-  };
-
-  if (iconData && iconData.frames.length > 0) {
-    console.info(`[MapCanvas] Pokémon icon lookup (${logContext})`, {
-      species,
-      form,
-      palette,
-      hasEntry: true,
-      frameCount: iconData.frames.length,
-      frameDurationFrames: iconData.frameDurationFrames,
-    });
-
-    const frames: SpriteFrameRef[] = iconData.frames.map((frameRecord, index) =>
-      convertFrame(frameRecord, index),
-    );
-    const baseFrame = frames[0] ?? null;
-    const frameSet = baseFrame
-      ? {
-          framesByDirection: { [direction]: frames },
-          availableDirections: [direction],
-          defaultFrame: baseFrame,
-          defaultDirection: direction,
-        }
-      : null;
-    const frameDurationMs = Math.max(
-      MIN_POKEMON_ICON_FRAME_DURATION_MS,
-      iconData.frameDurationFrames * FRAME_DURATION_MS * POKEMON_ICON_FRAME_DURATION_SCALE,
-    );
-    return {
-      baseFrame,
-      frameSet,
-      frameDurationMs,
-    };
-  }
-
-  console.warn(`[MapCanvas] Falling back to placeholder icon (${logContext})`, {
-    species,
-    form,
-    palette,
-    hasEntry: Boolean(iconData?.frames.length),
-  });
-
-  if (!fallbackRecord) {
-    return {
-      baseFrame: null,
-      frameSet: null,
-      frameDurationMs: null,
-    };
-  }
-
-  const halfWidth = Math.round((fallbackRecord.width ?? 0) / 2);
-  const halfHeight = Math.round((fallbackRecord.height ?? 0) / 2);
-  const fallbackFrame: SpriteFrameRef = {
-    key: `${spriteKey}:${facingKey}:fallback`,
-    texture: fallbackRecord.texture,
-    offsetX: fallbackRecord.offsetX + halfWidth,
-    offsetY: fallbackRecord.offsetY + halfHeight,
-  };
-  const fallbackFrameSet: MovementFrameSet = {
-    framesByDirection: { [direction]: [fallbackFrame] },
-    availableDirections: [direction],
-    defaultFrame: fallbackFrame,
-    defaultDirection: direction,
-  };
-
-  return {
-    baseFrame: fallbackFrame,
-    frameSet: fallbackFrameSet,
-    frameDurationMs: null,
-  };
-}
-
-function createAxisPathAnimator(
-  summary: MovementSummary,
-  objectEntry: ObjectEventEntry,
-  frameSet: MovementFrameSet | null,
-): PathMovementAnimator | null {
-  if (!summary.path || summary.path.length === 0) {
-    return null;
-  }
-  const start = summary.startCell;
-  const path = summary.path;
-  const recordedSteps = summary.steps ?? [];
-  const startIndex = path.findIndex((cell) => cell.x === start.x && cell.y === start.y);
-  if (startIndex === -1) {
-    return null;
-  }
-  const positiveOffsets = path
-    .slice(startIndex + 1)
-    .map((cell) => ({ dx: cell.x - start.x, dy: cell.y - start.y }));
-  const negativeOffsets = path
-    .slice(0, startIndex)
-    .map((cell) => ({ dx: cell.x - start.x, dy: cell.y - start.y }))
-    .reverse();
-
-  if (positiveOffsets.length === 0 && negativeOffsets.length === 0) {
-    return null;
-  }
-
-  const moveDuration = Math.max(90, movementSpeedToStepDuration(summary.model.speed));
-  const idleDuration = Math.max(120, movementSpeedToIdleDuration(summary.model.speed) * 0.6);
-
-  const segments: MovementSegment[] = [];
-  let current: Offset = { dx: 0, dy: 0 };
-  let currentDirection: CardinalDirection =
-    frameSet?.defaultDirection ?? (summary.axis === "x" ? "right" : "down");
-  let lastStepIndex: number = recordedSteps.length > 0 ? recordedSteps[0].index : 0;
-  let nextStepCursor = 0;
-  let moveSegmentCount = 0;
-
-  const rngSeedBase =
-    ((objectEntry.index ?? 0) * 1103515245 + start.x * 1237 + start.y * 1999) >>> 0;
-  const rng = createSeededRandom(rngSeedBase);
-  const earlyTurnChance = 0.35;
-  const earlyTurnFloor = 0.5;
-
-  const claimStepIndex = (target: Offset, direction: CardinalDirection): number => {
-    if (recordedSteps.length === 0) {
-      const index = nextStepCursor;
-      nextStepCursor += 1;
-      return index;
-    }
-    const targetX = start.x + target.dx;
-    const targetY = start.y + target.dy;
-    for (let cursor = nextStepCursor; cursor < recordedSteps.length; cursor += 1) {
-      const step = recordedSteps[cursor];
-      if (step.to.x === targetX && step.to.y === targetY && step.direction === direction) {
-        nextStepCursor = cursor + 1;
-        return step.index;
-      }
-    }
-    const fallbackStep =
-      recordedSteps[Math.min(recordedSteps.length - 1, nextStepCursor)] ??
-      recordedSteps[recordedSteps.length - 1];
-    nextStepCursor = Math.min(recordedSteps.length, nextStepCursor + 1);
-    return fallbackStep ? fallbackStep.index : recordedSteps.length;
-  };
-
-  const pushMove = (target: Offset): void => {
-    if (target.dx === current.dx && target.dy === current.dy) {
-      return;
-    }
-    const direction = deltaToDirection(target.dx - current.dx, target.dy - current.dy);
-    if (!direction) {
-      return;
-    }
-    const stepIndex = claimStepIndex(target, direction);
-    segments.push({
-      type: "move",
-      from: { ...current },
-      to: { ...target },
-      direction,
-      durationMs: moveDuration,
-      stepIndex,
-    });
-    current = { ...target };
-    currentDirection = direction;
-    lastStepIndex = stepIndex;
-    moveSegmentCount += 1;
-  };
-
-  const pushIdle = (duration: number): void => {
-    if (!(duration > 0)) {
-      return;
-    }
-    segments.push({
-      type: "wait",
-      position: { ...current },
-      direction: currentDirection,
-      durationMs: duration,
-      stepIndex: lastStepIndex,
-    });
-  };
-
-  const planSequences = (offsets: Offset[]): Offset[][] => {
-    if (offsets.length === 0) {
-      return [];
-    }
-    const passes = offsets.length > 1 ? 2 : 1;
-    const sequences: Offset[][] = [];
-    for (let pass = 0; pass < passes; pass += 1) {
-      let length = offsets.length;
-      if (pass > 0 && offsets.length > 1 && rng() < earlyTurnChance) {
-        const minSteps = Math.max(1, Math.floor(offsets.length * earlyTurnFloor));
-        const maxSteps = Math.max(minSteps, offsets.length - 1);
-        if (maxSteps > minSteps) {
-          const span = maxSteps - minSteps + 1;
-          length = minSteps + Math.floor(rng() * span);
-        } else {
-          length = minSteps;
-        }
-      }
-      length = Math.max(1, Math.min(offsets.length, length));
-      sequences.push(offsets.slice(0, length));
-    }
-    return sequences;
-  };
-
-  const runSequence = (offsets: Offset[]): void => {
-    if (!offsets.length) {
-      pushIdle(idleDuration);
-      return;
-    }
-    for (const offset of offsets) {
-      pushMove(offset);
-      pushIdle(idleDuration);
-    }
-    for (let index = offsets.length - 2; index >= 0; index -= 1) {
-      const offset = offsets[index];
-      if (!offset) {
-        continue;
-      }
-      pushMove(offset);
-      pushIdle(idleDuration);
-    }
-    pushMove({ dx: 0, dy: 0 });
-    pushIdle(idleDuration);
-  };
-
-  const positiveSequences = planSequences(positiveOffsets);
-  const negativeSequences = planSequences(negativeOffsets);
-  const totalSequences = Math.max(positiveSequences.length, negativeSequences.length);
-
-  if (totalSequences === 0) {
-    return null;
-  }
-
-  for (let pass = 0; pass < totalSequences; pass += 1) {
-    const forward = positiveSequences[pass];
-    if (forward) {
-      runSequence(forward);
-    }
-    const backward = negativeSequences[pass];
-    if (backward) {
-      runSequence(backward);
-    }
-  }
-
-  if (!segments.length) {
-    return null;
-  }
-  const totalDuration = segments.reduce(
-    (total, segment) => total + Math.max(0, segment.durationMs),
-    0,
-  );
-  if (!(totalDuration > 0)) {
-    return null;
-  }
-  const stepCount = recordedSteps.length > 0 ? recordedSteps.length : moveSegmentCount;
-
-  return {
-    kind: "path",
-    segments,
-    totalDurationMs: totalDuration,
-    stepCount,
-  };
-}
-
-function createWanderAnimator(
-  summary: MovementSummary,
-  objectEntry: ObjectEventEntry,
-  frameSet: MovementFrameSet | null,
-): PathMovementAnimator | null {
-  const bounds = summary.bounds ?? {
-    left: summary.startCell.x,
-    right: summary.startCell.x,
-    top: summary.startCell.y,
-    bottom: summary.startCell.y,
-  };
-  const reachableCells = summary.reachable ?? null;
-  const reachableSet = reachableCells
-    ? new Set(reachableCells.map((cell) => `${cell.x},${cell.y}`))
-    : null;
-
-  const seedBase =
-    (objectEntry.index ?? 0) * 1103515245 + summary.startCell.x * 1237 + summary.startCell.y * 1999;
-  const rng = createSeededRandom(seedBase >>> 0);
-  const moveDuration = Math.max(90, movementSpeedToStepDuration(summary.model.speed));
-  const idleBase = Math.max(120, movementSpeedToIdleDuration(summary.model.speed));
-  const directionPool: CardinalDirection[] = ["down", "up", "left", "right"];
-
-  let current: Offset = { dx: 0, dy: 0 };
-  let currentDirection: CardinalDirection = frameSet?.defaultDirection ?? "down";
-  let lastStepIndex = 0;
-  let nextStepIndex = 0;
-  let moveSegmentCount = 0;
-
-  const maxExtentX = bounds.right - bounds.left;
-  const maxExtentY = bounds.bottom - bounds.top;
-  const halfLength = Math.max(3, Math.min(12, maxExtentX + maxExtentY + 4));
-
-  const forwardDirections: CardinalDirection[] = [];
-
-  for (let step = 0; step < halfLength; step += 1) {
-    const candidates: CardinalDirection[] = [];
-    for (const direction of directionPool) {
-      const delta = DIRECTION_DELTAS[direction];
-      const next = { dx: current.dx + delta.dx, dy: current.dy + delta.dy };
-      const absX = summary.startCell.x + next.dx;
-      const absY = summary.startCell.y + next.dy;
-      if (absX < bounds.left || absX > bounds.right || absY < bounds.top || absY > bounds.bottom) {
-        continue;
-      }
-      if (reachableSet && !reachableSet.has(`${absX},${absY}`)) {
-        continue;
-      }
-      candidates.push(direction);
-    }
-    if (!candidates.length) {
-      break;
-    }
-    const choice = candidates[Math.floor(rng() * candidates.length)];
-    forwardDirections.push(choice);
-    const delta = DIRECTION_DELTAS[choice];
-    current = { dx: current.dx + delta.dx, dy: current.dy + delta.dy };
-  }
-
-  if (!forwardDirections.length) {
-    return null;
-  }
-
-  const directions: CardinalDirection[] = [...forwardDirections];
-  for (let index = forwardDirections.length - 1; index >= 0; index -= 1) {
-    directions.push(oppositeDirection(forwardDirections[index]));
-  }
-
-  current = { dx: 0, dy: 0 };
-  const segments: MovementSegment[] = [];
-  for (const direction of directions) {
-    const delta = DIRECTION_DELTAS[direction];
-    const target = { dx: current.dx + delta.dx, dy: current.dy + delta.dy };
-    const stepIndex = nextStepIndex;
-    nextStepIndex += 1;
-    segments.push({
-      type: "move",
-      from: { ...current },
-      to: target,
-      direction,
-      durationMs: moveDuration,
-      stepIndex,
-    });
-    current = target;
-    currentDirection = direction;
-    lastStepIndex = stepIndex;
-    moveSegmentCount += 1;
-    const idleDuration = idleBase * (0.6 + rng() * 0.6);
-    segments.push({
-      type: "wait",
-      position: { ...current },
-      direction: currentDirection,
-      durationMs: idleDuration,
-      stepIndex: lastStepIndex,
-    });
-  }
-
-  if (current.dx !== 0 || current.dy !== 0) {
-    while (current.dx !== 0) {
-      const direction = current.dx > 0 ? "left" : "right";
-      const delta = DIRECTION_DELTAS[direction];
-      const target = { dx: current.dx + delta.dx, dy: current.dy + delta.dy };
-      const stepIndex = nextStepIndex;
-      nextStepIndex += 1;
-      segments.push({
-        type: "move",
-        from: { ...current },
-        to: target,
-        direction,
-        durationMs: moveDuration,
-        stepIndex,
-      });
-      current = target;
-      currentDirection = direction;
-      lastStepIndex = stepIndex;
-      moveSegmentCount += 1;
-    }
-    while (current.dy !== 0) {
-      const direction = current.dy > 0 ? "up" : "down";
-      const delta = DIRECTION_DELTAS[direction];
-      const target = { dx: current.dx + delta.dx, dy: current.dy + delta.dy };
-      const stepIndex = nextStepIndex;
-      nextStepIndex += 1;
-      segments.push({
-        type: "move",
-        from: { ...current },
-        to: target,
-        direction,
-        durationMs: moveDuration,
-        stepIndex,
-      });
-      current = target;
-      currentDirection = direction;
-      lastStepIndex = stepIndex;
-      moveSegmentCount += 1;
-    }
-    segments.push({
-      type: "wait",
-      position: { ...current },
-      direction: currentDirection,
-      durationMs: idleBase,
-      stepIndex: moveSegmentCount > 0 ? lastStepIndex : 0,
-    });
-  }
-
-  const totalDuration = segments.reduce(
-    (total, segment) => total + Math.max(0, segment.durationMs),
-    0,
-  );
-  if (!(totalDuration > 0)) {
-    return null;
-  }
-  const stepCount = moveSegmentCount;
-  if (stepCount === 0) {
-    return null;
-  }
-
-  return {
-    kind: "path",
-    segments,
-    totalDurationMs: totalDuration,
-    stepCount,
-  };
-}
-
-function createSpinAnimator(
-  summary: MovementSummary,
-  objectEntry: ObjectEventEntry,
-  frameSet: MovementFrameSet | null,
-): SpinMovementAnimator | null {
-  const directionMode = summary.model.spinDirection ?? "random";
-  const interval = Math.max(180, movementSpeedToSpinInterval(summary.model.speed));
-  const steps: SpinStep[] = [];
-
-  if (directionMode === "clockwise" || directionMode === "counterclockwise") {
-    const sequence = directionMode === "clockwise" ? CLOCKWISE_SEQUENCE : COUNTERCLOCKWISE_SEQUENCE;
-    const startDirection = frameSet?.defaultDirection ?? sequence[0];
-    const startIndex = sequence.indexOf(startDirection);
-    const normalizedStart = startIndex >= 0 ? startIndex : 0;
-    for (let index = 0; index < sequence.length; index += 1) {
-      const direction = sequence[(normalizedStart + index) % sequence.length];
-      steps.push({ direction, durationMs: interval });
-    }
-  } else {
-    const available = frameSet?.availableDirections?.length
-      ? frameSet.availableDirections
-      : (["down", "up", "left", "right"] as CardinalDirection[]);
-    const rngSeed =
-      (objectEntry.index ?? 0) * 214013 +
-      summary.startCell.x * 2531011 +
-      summary.startCell.y * 1376312589;
-    const rng = createSeededRandom(rngSeed >>> 0);
-    const stepCount = 6;
-    for (let index = 0; index < stepCount; index += 1) {
-      const direction = available[Math.floor(rng() * available.length)] ?? "down";
-      const duration = interval * (0.6 + rng() * 0.8);
-      steps.push({ direction, durationMs: duration });
-    }
-  }
-
-  const totalDuration = steps.reduce((total, step) => total + Math.max(0, step.durationMs), 0);
-  if (!(totalDuration > 0)) {
-    return null;
-  }
-
-  return {
-    kind: "spin",
-    steps,
-    totalDurationMs: totalDuration,
-  };
-}
-
-function createIdleAnimator(
-  summary: MovementSummary,
-  objectEntry: ObjectEventEntry,
-  frameSet: MovementFrameSet | null,
-): IdleMovementAnimator | null {
-  if (!frameSet) {
-    return null;
-  }
-  const primaryDirection = frameSet.defaultDirection ?? frameSet.availableDirections[0] ?? null;
-  if (!primaryDirection) {
-    return null;
-  }
-  const frames = frameSet.framesByDirection[primaryDirection];
-  if (!frames || frames.length <= 1) {
-    return null;
-  }
-  const frameDuration = Math.max(260, movementSpeedToIdleDuration(summary.model.speed));
-  const seed =
-    ((objectEntry.index ?? 0) * 2147483647 +
-      summary.startCell.x * 2654435761 +
-      summary.startCell.y * 40503) >>>
-    0;
-  const rng = createSeededRandom(seed);
-  const phaseOffsetMs = Math.floor(rng() * frameDuration * frames.length);
-  return {
-    kind: "idle",
-    direction: primaryDirection,
-    frameCount: frames.length,
-    frameDurationMs: frameDuration,
-    phaseOffsetMs,
-  };
-}
-
-function createPokemonIconAnimator(
-  objectEntry: ObjectEventEntry,
-  frameSet: MovementFrameSet | null,
-  frameDurationMs: number,
-): IdleMovementAnimator | null {
-  if (!frameSet || !frameSet.defaultDirection) {
-    return null;
-  }
-  const direction = frameSet.defaultDirection;
-  const frames = frameSet.framesByDirection[direction];
-  if (!frames || frames.length <= 1) {
-    return null;
-  }
-  const duration = Math.max(60, Math.round(frameDurationMs));
-  const seed = ((objectEntry.index ?? 0) * 1103515245 + duration * 1664525) >>> 0;
-  const rng = createSeededRandom(seed);
-  const phaseOffsetMs = Math.floor(rng() * duration * frames.length);
-  return {
-    kind: "idle",
-    direction,
-    frameCount: frames.length,
-    frameDurationMs: duration,
-    phaseOffsetMs,
-  };
-}
-
-function createMovementAnimator(
-  summary: MovementSummary | null,
-  objectEntry: ObjectEventEntry,
-  frameSet: MovementFrameSet | null,
-): MovementAnimator | null {
-  if (!summary) {
-    return null;
-  }
-  if (summary.model.category === "axis-walk") {
-    return createAxisPathAnimator(summary, objectEntry, frameSet);
-  }
-  if (summary.model.category === "random-walk") {
-    return createWanderAnimator(summary, objectEntry, frameSet);
-  }
-  if (summary.model.category === "spin") {
-    return createSpinAnimator(summary, objectEntry, frameSet);
-  }
-  if (
-    summary.model.category === "object" &&
-    objectEntry.movement?.constant === "SPRITEMOVEDATA_BIG_GYARADOS"
-  ) {
-    const idleAnimator = createIdleAnimator(summary, objectEntry, frameSet);
-    if (idleAnimator) {
-      return idleAnimator;
-    }
-  }
-  return null;
-}
-
-function applySpriteFrame(
-  marker: ObjectMarkerEntry,
-  direction: CardinalDirection | null,
-  frameIndex: number,
-): void {
-  const frameSet = marker.frameSet;
-  if (!frameSet) {
-    if (direction) {
-      marker.lastDirection = direction;
-    }
-    return;
-  }
-  let frames =
-    direction && frameSet.framesByDirection[direction]?.length
-      ? frameSet.framesByDirection[direction]
-      : undefined;
-  let resolvedDirection = direction;
-  if ((!frames || frames.length === 0) && frameSet.defaultDirection) {
-    const fallback = frameSet.framesByDirection[frameSet.defaultDirection];
-    if (fallback && fallback.length) {
-      frames = fallback;
-      resolvedDirection = frameSet.defaultDirection;
-    }
-  }
-  if ((!frames || frames.length === 0) && frameSet.availableDirections.length > 0) {
-    const fallbackDirection = frameSet.availableDirections[0];
-    const fallback = frameSet.framesByDirection[fallbackDirection];
-    if (fallback && fallback.length) {
-      frames = fallback;
-      resolvedDirection = fallbackDirection;
-    }
-  }
-  let frame: SpriteFrameRef | null = null;
-  if (frames && frames.length > 0) {
-    const normalizedIndex = Math.max(0, Math.floor(frameIndex)) % frames.length;
-    frame = frames[normalizedIndex];
-  }
-  if (!frame) {
-    frame = frameSet.defaultFrame;
-  }
-  if (marker.currentFrameKey !== frame.key) {
-    marker.sprite.texture = frame.texture;
-    const spriteName = marker.object.sprite.constant;
-    if (
-      import.meta.env?.DEV &&
-      spriteName &&
-      (spriteName === "SPRITE_SAILBOAT" ||
-        spriteName === "SPRITE_BIG_GYARADOS" ||
-        spriteName === "SPRITE_BIG_SNORLAX")
-    ) {
-      const tex = frame.texture;
-      console.info(
-        `[SpriteCache] ${spriteName} frame ${frame.key} resolved ${tex.width}x${tex.height} (offset=${frame.offsetX},${frame.offsetY})`,
-      );
-    }
-    marker.currentFrameKey = frame.key;
-  }
-  marker.spriteOffset.x = frame.offsetX * marker.spriteScale;
-  marker.spriteOffset.y = frame.offsetY * marker.spriteScale;
-  if (resolvedDirection) {
-    marker.lastDirection = resolvedDirection;
-  }
-}
-
-function updateMarkerAnimation(marker: ObjectMarkerEntry, elapsedMs: number): void {
-  const baseX = marker.basePosition.x;
-  const baseY = marker.basePosition.y;
-  const cellSize = marker.cellPixelSize;
-  const animator = marker.animator;
-
-  if (!animator) {
-    applySpriteFrame(marker, marker.lastDirection ?? marker.frameSet?.defaultDirection ?? null, 0);
-    marker.sprite.x = baseX + marker.spriteOffset.x;
-    marker.sprite.y = baseY + marker.spriteOffset.y;
-    marker.stepCount = null;
-    marker.currentStepIndex = null;
-    marker.stepProgress = 0;
-    return;
-  }
-
-  if (animator.kind === "path") {
-    const total = animator.totalDurationMs;
-    marker.stepCount = animator.stepCount;
-    if (!(total > 0) || animator.segments.length === 0) {
-      applySpriteFrame(
-        marker,
-        marker.lastDirection ?? marker.frameSet?.defaultDirection ?? null,
-        0,
-      );
-      marker.sprite.x = baseX + marker.spriteOffset.x;
-      marker.sprite.y = baseY + marker.spriteOffset.y;
-      marker.currentStepIndex = null;
-      marker.stepProgress = 0;
-      return;
-    }
-    const timeInCycle = ((elapsedMs % total) + total) % total;
-    let accumulator = 0;
-    let activeSegment = animator.segments[animator.segments.length - 1];
-    for (const segment of animator.segments) {
-      const next = accumulator + segment.durationMs;
-      if (timeInCycle < next) {
-        activeSegment = segment;
-        break;
-      }
-      accumulator = next;
-    }
-    if (activeSegment.type === "move") {
-      const segmentElapsed = timeInCycle - accumulator;
-      const duration = activeSegment.durationMs > 0 ? activeSegment.durationMs : 1;
-      const progress = Math.max(0, Math.min(1, segmentElapsed / duration));
-      const interpDx =
-        activeSegment.from.dx + (activeSegment.to.dx - activeSegment.from.dx) * progress;
-      const interpDy =
-        activeSegment.from.dy + (activeSegment.to.dy - activeSegment.from.dy) * progress;
-      const stepIndex = activeSegment.stepIndex ?? 0;
-      const parity = stepIndex & 1;
-      const inStride = progress >= 0.5;
-      const baseFrameIndex = parity === 0 ? 0 : 2;
-      const strideFrameIndex = parity === 0 ? 1 : 3;
-      const frameIndex = inStride ? strideFrameIndex : baseFrameIndex;
-      applySpriteFrame(marker, activeSegment.direction, frameIndex);
-      marker.sprite.x = baseX + marker.spriteOffset.x + interpDx * cellSize;
-      marker.sprite.y = baseY + marker.spriteOffset.y + interpDy * cellSize;
-      marker.currentStepIndex = activeSegment.stepIndex;
-      marker.stepProgress = progress;
-    } else {
-      const stepIndex = activeSegment.stepIndex ?? 0;
-      const parity = stepIndex & 1;
-      const frameIndex = parity === 0 ? 0 : 2;
-      applySpriteFrame(marker, activeSegment.direction, frameIndex);
-      marker.sprite.x = baseX + marker.spriteOffset.x + activeSegment.position.dx * cellSize;
-      marker.sprite.y = baseY + marker.spriteOffset.y + activeSegment.position.dy * cellSize;
-      marker.currentStepIndex = activeSegment.stepIndex;
-      marker.stepProgress = 0;
-    }
-    return;
-  }
-
-  if (animator.kind === "idle") {
-    const frameCount = Math.max(1, animator.frameCount);
-    const frameDuration = Math.max(1, animator.frameDurationMs);
-    const loopDuration = frameCount * frameDuration;
-    const cycleTime =
-      (((elapsedMs + animator.phaseOffsetMs) % loopDuration) + loopDuration) % loopDuration;
-    const frameIndex = Math.floor(cycleTime / frameDuration) % frameCount;
-    applySpriteFrame(marker, animator.direction, frameIndex);
-    marker.sprite.x = baseX + marker.spriteOffset.x;
-    marker.sprite.y = baseY + marker.spriteOffset.y;
-    marker.lastDirection = animator.direction ?? marker.lastDirection;
-    marker.stepCount = null;
-    marker.currentStepIndex = null;
-    marker.stepProgress = 0;
-    return;
-  }
-
-  if (animator.kind === "spin") {
-    const total = animator.totalDurationMs;
-    if (!(total > 0) || animator.steps.length === 0) {
-      applySpriteFrame(
-        marker,
-        marker.lastDirection ?? marker.frameSet?.defaultDirection ?? null,
-        0,
-      );
-      marker.sprite.x = baseX + marker.spriteOffset.x;
-      marker.sprite.y = baseY + marker.spriteOffset.y;
-      marker.stepCount = null;
-      marker.currentStepIndex = null;
-      marker.stepProgress = 0;
-      return;
-    }
-    const timeInCycle = ((elapsedMs % total) + total) % total;
-    let accumulator = 0;
-    let activeStep = animator.steps[animator.steps.length - 1];
-    for (const step of animator.steps) {
-      const next = accumulator + step.durationMs;
-      if (timeInCycle < next) {
-        activeStep = step;
-        break;
-      }
-      accumulator = next;
-    }
-    applySpriteFrame(marker, activeStep.direction, 0);
-    marker.sprite.x = baseX + marker.spriteOffset.x;
-    marker.sprite.y = baseY + marker.spriteOffset.y;
-    marker.lastDirection = activeStep.direction;
-    marker.stepCount = null;
-    marker.currentStepIndex = null;
-    marker.stepProgress = 0;
-    return;
-  }
-
-  applySpriteFrame(marker, marker.lastDirection ?? marker.frameSet?.defaultDirection ?? null, 0);
-  marker.sprite.x = baseX + marker.spriteOffset.x;
-  marker.sprite.y = baseY + marker.spriteOffset.y;
-  marker.stepCount = null;
-  marker.currentStepIndex = null;
-  marker.stepProgress = 0;
-}
-
-function snapToHalf(value: number): number {
-  if (!Number.isFinite(value)) {
-    return 0;
-  }
-  return Math.round(value * 2) / 2;
-}
 
 export default function MapCanvas({
   atlas,
