+++ web/atlas/src/components/MapCanvas/index.ts
@@ -1,0 +1,3 @@
+// Re-export types, constants, and the main component
+export * from "./MapCanvas.types";
+export * from "./constants";
