+++ web/atlas/src/lib/map-canvas/index.ts
@@ -1,0 +1,5 @@
+// Re-export all map canvas utilities
+export * from "./viewport-utils";
+export * from "./animation-frames";
+export * from "./movement-animators";
+export * from "./sprite-animation";
